# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /generations/{id}:
    get:
      operationId: getGeneration
      tags: [generations]
      summary: Get a single generation with its full markdown body
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Generation"
        "404":
          description: Generation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGeneration
      tags: [generations]
      summary: Delete a generation from the archive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Generation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /generations/{id}/feedback:
    put:
      operationId: setGenerationFeedback
      tags: [generations]
      summary: Record the human QA verdict on a generation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "200":
          description: The updated generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Generation"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Generation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /generations/{id}/approve:
    post:
      operationId: approveGeneration
      tags: [generations]
      summary: Approve a held client-facing report and send it to the client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The updated generation (report sent)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Generation"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Generation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: No report is awaiting approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: The held report payload is unreadable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Sending the report failed (draft stays pending)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /generations/{id}/request-changes:
    post:
      operationId: requestGenerationChanges
      tags: [generations]
      summary: Hold a report back for rework with reviewer notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestChangesInput"
      responses:
        "200":
          description: The updated generation (held back)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Generation"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Generation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: No report is awaiting approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /generations/{id}/steps:
    get:
      operationId: getGenerationSteps
      tags: [generations]
      summary: Per-agent audit-trail steps for a generation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The generation's steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationStepList"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /generations/{id}/proposals:
    get:
      operationId: getProposals
      tags: [proposals]
      summary: List improvement proposals for a generation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The proposals for this generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProposalList"
    post:
      operationId: createProposals
      tags: [proposals]
      summary: Generate improvement proposals from the generation's feedback
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The generated proposals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProposalList"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Generation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Proposal generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Generation:
      type: object
      required:
        - id
        - clientName
        - workflowTitle
        - leadAgentTitle
        - teamTitles
        - requestText
        - createdAt
        - clientPath
        - workflowPath
        - leadAgentPath
        - teamPaths
        - finalMarkdown
        - status
        - triggerSource
        - durationMs
        - totalTokens
        - feedbackVerdict
        - feedbackNote
        - feedbackAt
        - approvalStatus
        - approvalNote
        - approvalAt
        - hasPendingDelivery
        - pendingDeliveryKind
        - pendingEmailReply
      properties:
        id:
          type: integer
        clientName:
          type: string
        workflowTitle:
          type: string
        leadAgentTitle:
          type: string
        teamTitles:
          type: array
          items:
            type: string
        requestText:
          type: string
        createdAt:
          type: string
          format: date-time
        clientPath:
          type: string
        workflowPath:
          type: string
        leadAgentPath:
          type: string
        teamPaths:
          type: array
          items:
            type: string
        finalMarkdown:
          type: string
        status:
          type: string
        triggerSource:
          type: string
        durationMs:
          type: ["integer", "null"]
        totalTokens:
          type: ["integer", "null"]
        feedbackVerdict:
          type: ["string", "null"]
        feedbackNote:
          type: ["string", "null"]
        feedbackAt:
          type: ["string", "null"]
          format: date-time
        approvalStatus:
          type: ["string", "null"]
        approvalNote:
          type: ["string", "null"]
        approvalAt:
          type: ["string", "null"]
          format: date-time
        hasPendingDelivery:
          type: boolean
        pendingDeliveryKind:
          type: ["string", "null"]
          enum: ["monthly-report", "email-reply", null]
          description: >-
            Which kind of drafted-but-unsent delivery is held on this run, or
            null when none is pending.
        pendingEmailReply:
          type: ["object", "null"]
          description: >-
            For a held inbound email reply (drafted autonomously, no live
            session), the reviewable content so a human can approve from the
            archive. Identity/threading internals are intentionally omitted.
          required:
            - recipient
            - subject
            - inboundText
            - replyBody
          properties:
            recipient:
              type: string
            subject:
              type: string
            inboundText:
              type: string
            replyBody:
              type: string
...
    GenerationStep:
      type: object
      required:
        - id
        - agentPath
        - agentTitle
        - stepOrder
        - role
        - status
        - durationMs
        - inputTokens
        - outputTokens
        - charCount
        - errorMessage
        - createdAt
      properties:
        id:
          type: integer
        agentPath:
          type: string
        agentTitle:
          type: string
        stepOrder:
          type: integer
        role:
          type: string
        status:
          type: string
        durationMs:
          type: ["integer", "null"]
        inputTokens:
          type: ["integer", "null"]
        outputTokens:
          type: ["integer", "null"]
        charCount:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    GenerationStepList:
      type: object
      required: [steps]
      properties:
        steps:
          type: array
          items:
            $ref: "#/components/schemas/GenerationStep"
...
    FeedbackInput:
      type: object
      required: [verdict]
      properties:
        verdict:
          type: string
          enum: [approved, rejected]
        note:
          type: ["string", "null"]
    RequestChangesInput:
      type: object
      properties:
        note:
          type: ["string", "null"]
    ImprovementProposal:
      type: object
      required:
        - id
        - generationId
        - targetType
        - targetPath
        - targetLabel
        - rationale
        - proposedText
        - status
        - createdAt
        - decidedAt
      properties:
        id:
          type: integer
        generationId:
          type: integer
        targetType:
          type: string
          description: "knowledge (an agency-wide standard) or client (this client only)."
        targetPath:
          type: string
        targetLabel:
          type: string
        rationale:
          type: string
        proposedText:
          type: string
        status:
          type: string
          description: "pending, accepted, or rejected."
        createdAt:
          type: string
          format: date-time
        decidedAt:
          type: ["string", "null"]
          format: date-time
    ProposalList:
      type: object
      required: [proposals]
      properties:
        proposals:
          type: array
          items:
            $ref: "#/components/schemas/ImprovementProposal"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string