# lib/api-zod/src/generated/types/generation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GenerationPendingDeliveryKind } from './generationPendingDeliveryKind';
import type { GenerationPendingEmailReply } from './generationPendingEmailReply';

export interface Generation {
  id: number;
  clientName: string;
  workflowTitle: string;
  leadAgentTitle: string;
  teamTitles: string[];
  requestText: string;
  createdAt: Date;
  clientPath: string;
  workflowPath: string;
  leadAgentPath: string;
  teamPaths: string[];
  finalMarkdown: string;
  status: string;
  triggerSource: string;
  /** @nullable */
  durationMs: number | null;
  /** @nullable */
  totalTokens: number | null;
  /** @nullable */
  feedbackVerdict: string | null;
  /** @nullable */
  feedbackNote: string | null;
  /** @nullable */
  feedbackAt: Date | null;
  /** @nullable */
  approvalStatus: string | null;
  /** @nullable */
  approvalNote: string | null;
  /** @nullable */
  approvalAt: Date | null;
  hasPendingDelivery: boolean;
  /**
     * Which kind of drafted-but-unsent delivery is held on this run, or null when none is pending.
     * @nullable
     */
  pendingDeliveryKind: GenerationPendingDeliveryKind;
  /**
     * For a held inbound email reply (drafted autonomously, no live session), the reviewable content so a human can approve from the archive. Identity/threading internals are intentionally omitted.
     * @nullable
     */
  pendingEmailReply: GenerationPendingEmailReply;
}